+++ docs/guides/changelog.md
@@ -17,6 +17,7 @@ description: |-
     - `set_bgp_*_next_hop_specific_ip` to `set_bgp_*_next_hop_specific_ips`
 - BREAKING CHANGE: `fmc_ikev1_policies`: Rename attribute `encryption` to `encryption_algorithm`
 - BREAKING CHANGE: `fmc_certificate_enrollment`: Rename attribute `crl_static_urls_list` to `crl_static_urls`
+- BREAKING CHANGE: `fmc_device_vti_interface`: Rename attribute `ipv4_address/netmask` to `ipv4_static_address/netmask`
 - (Fix) Updated `required` flag for several attributs in multiple resources
 - (Fix) `fmc_chassis_logical_device`: Deletion would fail if registered device is under deployment
 - (Enhancement) Add multiple bulk/non-bulk variants of already existing resources and data sources